mktemp -t requires a template on linux
mktemp -t on modern coreutils says 'template' which requires a block of XXX
to be templatized. On mac -t does not use a template. This means that on linux
we'll get a file like    keys-deadbeef and on mac we'll get keys-XXXXXXXX-deadbeef
but who cares, it's a random file either way. And now it actually works on linux.