manifests.rhel/0000_90_openshift-cluster-image-policy: Drop TechPreviewNoUpgrade limitation
Now that ClusterImagePolicy is v1 and in the default feature set, we
can remove the annotation that limites the openshift
ClusterImagePolicy to the TechPreviewNoUpgrade feature set.  No CI
impact is expected, because even 4.y.z release-controller jobs [1]
like 4.19.8 -> 4.19.9 [2] use imported-into-CI-registry versions of
the target release like registry.ci.openshift.org/ocp/release:4.19.9
[3]. But it sets us up for being able to drop OpenPGP/GPG signature
verification in future releases, and avoids the risk that a 4.20
customer creates their own openshift ClusterImagePolicy and the
resulting de-conflicting guards we'd need on 4.20-to-4.21 updates.

[1]: https://amd64.ocp.releases.ci.openshift.org/releasestream/4-stable/release/4.19.9
[2]: https://prow.ci.openshift.org/view/gs/test-platform-results/logs/release-openshift-origin-installer-e2e-aws-upgrade/1956054636976672768
[3]: https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/test-platform-results/logs/release-openshift-origin-installer-e2e-aws-upgrade/1956054636976672768/artifacts/e2e-aws-upgrade/clusterversion.json